refactor: ui 패키지 controller 로 이름 변경
