+++ src/main/java/wooteco/subway/controller/SubwayControllerAdvice.java
@@ -1,4 +1,4 @@
-package wooteco.subway.ui;
+package wooteco.subway.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.servlet.NoHandlerFoundException;
-import wooteco.subway.ui.dto.ErrorResponse;
+import wooteco.subway.controller.dto.ErrorResponse;
 
 @ControllerAdvice
 public class SubwayControllerAdvice {
